Amend file names
