+++ __test__/setupTest.js
@@ -18,8 +18,9 @@ beforeAll(async () => {
   app.use((req, res, next) => {
     req.user = {
       _id: "user_id",
-      name: "testuser",
+      name: "test_user",
       email: "testuser@gmail.com",
+      password:"123"
     };
     next();
   });
@@ -39,3 +40,32 @@ module.exports ={
     app,
     generateToken
 };
+
+// const {MongoClient} = require('mongodb');
+
+// describe('insert', () => {
+//   let connection;
+//   let db;
+
+//   beforeAll(async () => {
+//     connection = await MongoClient.connect(globalThis.__MONGO_URI__, {
+//       useNewUrlParser: true,
+//       useUnifiedTopology: true,
+//     });
+//     db = await connection.db(globalThis.__MONGO_DB_NAME__);
+//   });
+
+//   afterAll(async () => {
+//     await connection.close();
+//   });
+
+//   it('should insert a doc into collection', async () => {
+//     const users = db.collection('users');
+
+//     const mockUser = {_id: 'some-user-id', name: 'John'};
+//     await users.insertOne(mockUser);
+
+//     const insertedUser = await users.findOne({_id: 'some-user-id'});
+//     expect(insertedUser).toEqual(mockUser);
+//   });
+// });
