+++ config/db-setup.js
@@ -2,10 +2,15 @@ require('dotenv').config()
 const mongoose =require( "mongoose");
 
 const connectionUrl = process.env.DB_CONNECTION
+const dbName = process.env.DB_NAME
 
 const connectDB = async () => {
     try {
-       const conn = await mongoose.connect(connectionUrl);
+       const conn = await mongoose.connect(connectionUrl,{
+        dbName,
+        useNewUrlParser: true,     
+        useUnifiedTopology: true,  
+      });
         console.log(` MongoDB Connected:${conn.connection.host} 🚀`)
     } catch (error) {
         console.error(`Error:${error.message}`);
