+++ controllers/Users.js
@@ -1,5 +1,5 @@
 const asyncHandler = require ('express-async-handler');
-const User = require ('../models/userModels.js');
+const User = require ('../models/Users.js');
 const generateToken = require ('../utils/generateToken.js');
 
 // @desc    Auth user & get token
