+++ jest.config.js
@@ -1,4 +1,3 @@
-module.exports =
-{
-    "preset": "@shelf/jest-mongodb"
-  }
+module.exports = {
+  preset: "@shelf/jest-mongodb",
+};
